Move viewsets to own modules
