+++ cms/topics/viewsets/__init__.py
@@ -1,0 +1,3 @@
+from .featured_series import *  # noqa: F403
+from .highlighted_items import *  # noqa: F403
+from .series_with_headline_figures import *  # noqa: F403
